refactor(queue): extract duplicate-cluster adjudication into its own module
Part of #4013's module-split sequence (step 3, after transient-locks.ts
and signal-snapshot.ts): dupWinnerLinkedDuplicateCount,
dupWinnerLinkedDuplicateWinnerNumber, reconcileLiveDuplicateSiblings,
linkedIssueDuplicatePullRequestsForGate, and
linkedIssueDuplicatePullRequestRecordsForGate move to
src/queue/duplicate-detection.ts. Pure mechanical move, no behavior
change -- a re-export shim keeps processors.ts's own internal callers
and the existing test/unit/duplicate-winner.test.ts and
test/unit/reconcile-live-duplicate-siblings.test.ts imports working
unchanged.
